+++ examples/js-demo/app.js
@@ -1,5 +1,6 @@
 /** @jsx h */
 import { autocomplete, getAlgoliaResults } from '@algolia/autocomplete-js';
+import algoliarecommend from '@algolia/recommend';
 import {
   frequentlyBoughtTogether,
   relatedProducts,
@@ -19,6 +20,7 @@ const apiKey = '28cf6d38411215e2eef188e635216508';
 const indexName = 'gstar_demo_test';
 
 const searchClient = algoliasearch(appId, apiKey);
+const recommendClient = algoliarecommend(appId, apiKey);
 
 insights('init', { appId, apiKey });
 insights('setUserToken', 'user-token-1');
@@ -140,7 +142,7 @@ function ReferenceItem({ item }) {
 function renderRecommendations(selectedProduct) {
   frequentlyBoughtTogether({
     container: '#frequentlyBoughtTogether',
-    searchClient,
+    recommendClient,
     indexName,
     objectIDs: [selectedProduct.objectID],
     itemComponent({ item }) {
@@ -153,13 +155,13 @@ function renderRecommendations(selectedProduct) {
       );
     },
     maxRecommendations: 3,
-    searchParameters: {
+    queryParameters: {
       analytics: true,
       clickAnalytics: true,
     },
     fallbackComponent() {
       return relatedProducts({
-        searchClient,
+        recommendClient,
         indexName,
         objectIDs: [selectedProduct.objectID],
         itemComponent({ item }) {
@@ -176,10 +178,12 @@ function renderRecommendations(selectedProduct) {
         translations: {
           title: 'Related products (fallback)',
         },
-        fallbackFilters: [
-          `hierarchical_categories.lvl2:${selectedProduct.hierarchical_categories.lvl2}`,
-        ],
-        searchParameters: {
+        fallbackParameters: {
+          facetFilters: [
+            `hierarchical_categories.lvl2:${selectedProduct.hierarchical_categories.lvl2}`,
+          ],
+        },
+        queryParameters: {
           analytics: true,
           clickAnalytics: true,
           facetFilters: [
@@ -192,7 +196,7 @@ function renderRecommendations(selectedProduct) {
 
   relatedProducts({
     container: '#relatedProducts',
-    searchClient,
+    recommendClient,
     indexName,
     objectIDs: [selectedProduct.objectID],
     itemComponent({ item }) {
@@ -209,10 +213,12 @@ function renderRecommendations(selectedProduct) {
     translations: {
       title: 'Related products',
     },
-    fallbackFilters: [
-      `hierarchical_categories.lvl2:${selectedProduct.hierarchical_categories.lvl2}`,
-    ],
-    searchParameters: {
+    fallbackParameters: {
+      facetFilters: [
+        `hierarchical_categories.lvl2:${selectedProduct.hierarchical_categories.lvl2}`,
+      ],
+    },
+    queryParameters: {
       analytics: true,
       clickAnalytics: true,
       facetFilters: [