+++ packages/recommend-core/src/types/ProductRecord.ts
@@ -1,7 +1,4 @@
 export type ProductRecord<TObject> = TObject & {
   objectID: string;
-  __indexName: string;
-  __queryID: string | undefined;
-  __position: number;
-  __recommendScore: number | null;
+  _score?: number;
 };