+++ packages/recommend-core/src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './ProductRecord';
 export * from './RecommendModel';
-export * from './RecommendRecord';
 export * from './RecordWithObjectID';