+++ packages/recommend-core/src/utils/index.ts
@@ -1,6 +1,3 @@
-export * from './cx';
-export * from './getHitsPerPage';
-export * from './getIndexNameFromModel';
-export * from './getSearchParametersForModel';
+export * from './mapToRecommendations';
 export * from './sortBy';
 export * from './uniqBy';