+++ packages/recommend-react/src/FrequentlyBoughtTogether.tsx
@@ -1,6 +1,7 @@
+import { GetFrequentlyBoughtTogetherProps } from '@algolia/recommend-core';
 import {
   createFrequentlyBoughtTogetherComponent,
-  FrequentlyBoughtTogetherProps,
+  FrequentlyBoughtTogetherProps as FrequentlyBoughtTogetherVDOMProps,
 } from '@algolia/recommend-vdom';
 import React, { createElement, Fragment } from 'react';
 
@@ -13,12 +14,23 @@ const UncontrolledFrequentlyBoughtTogether = createFrequentlyBoughtTogetherCompo
   }
 );
 
+type FrequentlyBoughtTogetherProps<
+  TObject
+> = GetFrequentlyBoughtTogetherProps<TObject> &
+  Omit<FrequentlyBoughtTogetherVDOMProps<TObject>, 'items' | 'status'>;
+
 export function FrequentlyBoughtTogether<TObject>(
   props: FrequentlyBoughtTogetherProps<TObject>
 ) {
-  const { recommendations } = useFrequentlyBoughtTogether<TObject>(props);
+  const { recommendations, status } = useFrequentlyBoughtTogether<TObject>(
+    props
+  );
 
   return (
-    <UncontrolledFrequentlyBoughtTogether {...props} items={recommendations} />
+    <UncontrolledFrequentlyBoughtTogether
+      {...props}
+      items={recommendations}
+      status={status}
+    />
   );
 }